requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
cycler==0.10.0
idna==2.5
jieba==0.38
matplotlib==2.0.2
numpy==1.13.1
olefile==0.44
Pillow==4.2.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.3
scipy==0.19.1
six==1.10.0
urllib3==1.22
wordcloud==1.3.1
